This version of MiniCache supports SHFS 1.06,
    comes with a telnet shell, and an http server
    that directly operates on top of SHFS. Cache
    actions can be triggered via XenStore and
    download statistics are collected for each
    object on hits and misses. It uses the upstream
    version of Mini-OS from cnplab. The upstream
    toolchain is also required.
    x86_64 and ARM32 builds are supported.

    Please consider this release as
    experimental.

+++ blkdev.h
@@ -111,7 +111,7 @@ static inline int blkdev_async_io(struct blkdev *bd, sector_t start, sector_t le
 		return -ENXIO;
 	}
 
-	if (unlikely(((uint64_t) buffer) & ((uint64_t) blkdev_ssize(bd) - 1))) {
+	if (unlikely(((uintptr_t) buffer) & ((uintptr_t) blkdev_ssize(bd) - 1))) {
 		/* buffer is not aligned to device sector size */
 		return -EINVAL;
 	}
